[fix] todoDAO에서 Int -> Long으로 변경

+++ DontForgetProject/app/src/main/java/com/test/dontforgetproject/DAO/TodoDAO.kt
@@ -1,15 +1,15 @@
 package com.test.dontforgetproject.DAO
 
 data class TodoClass(
-    var todoIdx: Int,
+    var todoIdx: Long,
     var todoContent: String,
-    var todoIsChecked: Int,
-    var todoCategoryIdx: Int,
+    var todoIsChecked: Long,
+    var todoCategoryIdx: Long,
     var todoCategoryName: String,
     var todoDate: String,
     var todoAlertTime: String,
     var todoLocationLatitude: String,
     var todoLocationLongitude: String,
-    var todoOwnerIdx: Int,
+    var todoOwnerIdx: Long,
     var todoOwnerName: String
 )
